# trackingservice/target/classes/openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: contact@egovernments.org
  description: "Tracking service provides client application the capability to create\
    \ routes, trips and track the trips. This service is also responsible for generating\
    \ alerts and notifications in case of special event."
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://sanitation.digit.org/
  title: Tracking service
  version: 1.0.11
...
servers:
- url: https://core.digit.org/api/v3
...
paths:
  /poi/_update:
    put:
      description: Update an existing POI by Id
      operationId: updatePOI
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/POI'
          application/xml:
            schema:
              $ref: '#/components/schemas/POI'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/POI'
        description: Update an existent POI in the system
        required: true
      responses:
        "200":
          description: Successful operation
        "400":
          description: Invalid POI ID is supplied
        "404":
          description: POI not found
        "405":
          description: Validation exception
      summary: Update an existing POI using its id
      tags:
      - POI
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: POI
  /poi/_create:
    post:
      description: Create a new POI by Id
      operationId: createPOI
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/POI'
          application/xml:
            schema:
              $ref: '#/components/schemas/POI'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/POI'
        description: Create a new POI in the system
        required: true
      responses:
        "200":
          description: Successful operation
        "405":
          description: Validation exception
      summary: Create a new POI by providing location coordinates and additional details
      tags:
      - POI
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: POI
  /poi/_search:
    get:
      description: Search POIs based on multiple filters
      operationId: findPOI
      parameters:
      - description: Status values that need to be considered for filter
        explode: true
        in: query
        name: status
        required: false
        schema:
          default: active
          enum:
          - active
          - inactive
          type: string
        style: form
      - description: Location name that needs to be considered for filter
        explode: true
        in: query
        name: locationName
        required: false
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/POI'
                type: array
            application/xml:
              schema:
                items:
                  $ref: '#/components/schemas/POI'
                type: array
          description: successful operation
        "400":
          description: Invalid search value
      summary: Finds POIs based on a specific parameters
      tags:
      - POI
      x-accepts: application/json
      x-tags:
      - tag: POI
  /poi/{poiId}:
    get:
      description: Returns a single POI
      operationId: getPoiById
      parameters:
      - description: ID of POI to return
        explode: false
        in: path
        name: poiId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/POI'
            application/xml:
              schema:
                $ref: '#/components/schemas/POI'
          description: successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: POI not found
      summary: Find POI by its id
      tags:
      - POI
      x-accepts: application/json
      x-tags:
      - tag: POI
...
components:
  schemas:
    POI:
      example:
        locationName: Any name assigned to the location
        alert:
        - alert
        - alert
        audit:
          createdDate: 2000-01-23T04:56:07.000+00:00
          updatedBy: Id of the user who updated the entity
          createdBy: Id of the user who created the entity
          updatedDate: 2000-01-23T04:56:07.000+00:00
        locationDetails:
        - latitude: 0.8008282
          longitude: 6.0274563
        - latitude: 0.8008282
          longitude: 6.0274563
        id: id
        type: point
        status: active
      properties:
        id:
          description: GUID generated during POI creation
          type: string
        locationName:
          example: Any name assigned to the location
          type: string
        status:
          description: POI status to help to exclude or include this in route planning
          enum:
          - active
          - inactive
          example: active
          type: string
        type:
          description: POI type helps identify the shape of the location
          enum:
          - point
          - line
          - polygon
          example: point
          type: string
        locationDetails:
          description: A POI can comprise of on geocoordinate in case of single point
            and multiple geos in case of a polygon
          items:
            $ref: '#/components/schemas/Location'
          type: array
          xml:
            wrapped: true
        alert:
          description: List of alert codes that may be applicable to the location.
            This is optional
          items:
            type: string
          type: array
        audit:
          $ref: '#/components/schemas/Audit'
      type: object
      xml:
        name: poi
    Location:
      example:
        latitude: 0.8008282
        longitude: 6.0274563
      properties:
        latitude:
          format: float
          type: number
        longitude:
          format: float
          type: number
      type: object
      xml:
        name: location
    Audit:
      example:
        createdDate: 2000-01-23T04:56:07.000+00:00
        updatedBy: Id of the user who updated the entity
        createdBy: Id of the user who created the entity
        updatedDate: 2000-01-23T04:56:07.000+00:00
      properties:
        createdBy:
          example: Id of the user who created the entity
          type: string
        createdDate:
          format: date-time
          type: string
        updatedBy:
          example: Id of the user who updated the entity
          type: string
        updatedDate:
          format: date-time
          type: string
      type: object
      xml:
        name: audit
...